Handling some nullpointer exception.

+++ src/main/java/de/reza/documentclassifier/utils/DatasetProcessor.java
@@ -9,7 +9,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-
 @Service
 @Slf4j
 public class DatasetProcessor {
@@ -20,25 +19,26 @@ public class DatasetProcessor {
      * @param uuid      uuid of the model
      * @return          Path to the pdf files.
      */
-    public String unzip(MultipartFile dataset, String uuid){
+    public String unzip(MultipartFile dataset, String uuid) {
 
-        File dic = new File("models/" + uuid);
-        boolean createFolderSuccessful = dic.mkdirs();
-        if(createFolderSuccessful) {
-            Path filepath = Paths.get("models/" + uuid, dataset.getOriginalFilename());
-            try {
-                dataset.transferTo(filepath);
-                new ZipFile(filepath.toFile()).extractAll("models/" + uuid + "/trainingFiles/");
-                boolean removeZipFile = filepath.toFile().delete();
-                if(removeZipFile){
-                    log.info("Zip file deleted");
-                    return "models/" + uuid + "/trainingFiles/";
+        if (!dataset.isEmpty()) {
+            File dic = new File("models/" + uuid);
+            boolean createFolderSuccessful = dic.mkdirs();
+            if (createFolderSuccessful) {
+                Path filepath = Paths.get("models/" + uuid, dataset.getOriginalFilename());
+                try {
+                    dataset.transferTo(filepath);
+                    new ZipFile(filepath.toFile()).extractAll("models/" + uuid + "/trainingFiles/");
+                    boolean removeZipFile = filepath.toFile().delete();
+                    if (removeZipFile) {
+                        log.info("Zip file deleted");
+                        return "models/" + uuid + "/trainingFiles/";
+                    }
+                } catch (IOException e) {
+                    return null;
                 }
-            } catch (IOException e) {
-                throw new RuntimeException(e);
             }
         }
-        return null;
+            return null;
     }
-
 }
